WFLY-18879 Consolidate CommandDispatcherFactory interfaces from API vs SPI modules
