+++ clustering/server/extension/src/main/java/org/wildfly/extension/clustering/server/dispatcher/LocalCommandDispatcherFactoryServiceConfigurator.java
@@ -15,8 +15,8 @@ import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceTarget;
+import org.wildfly.clustering.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.group.Group;
-import org.wildfly.clustering.server.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.server.infinispan.dispatcher.LocalCommandDispatcherFactory;
 import org.wildfly.clustering.server.service.ClusteringRequirement;
 import org.wildfly.clustering.service.FunctionalService;
