+++ clustering/server/extension/src/main/java/org/wildfly/extension/clustering/server/provider/CacheServiceProviderRegistryServiceConfigurator.java
@@ -17,9 +17,9 @@ import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceTarget;
+import org.wildfly.clustering.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.infinispan.service.InfinispanCacheRequirement;
 import org.wildfly.clustering.provider.ServiceProviderRegistry;
-import org.wildfly.clustering.server.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.server.group.Group;
 import org.wildfly.clustering.server.infinispan.provider.AutoCloseableServiceProviderRegistry;
 import org.wildfly.clustering.server.infinispan.provider.CacheServiceProviderRegistry;
