+++ clustering/server/service/src/main/java/org/wildfly/clustering/server/service/group/LocalCacheGroupServiceConfiguratorProvider.java
@@ -5,11 +5,10 @@
 
 package org.wildfly.clustering.server.service.group;
 
-import org.wildfly.clustering.server.group.Group;
 import org.wildfly.clustering.server.service.LocalCacheServiceConfiguratorProvider;
 
 /**
- * Provides service configurators for a local cache {@link Group}.
+ * Provides service configurators for a local cache {@link org.wildfly.clustering.group.Group}.
  * @author Paul Ferraro
  */
 public interface LocalCacheGroupServiceConfiguratorProvider extends LocalCacheServiceConfiguratorProvider {
