+++ clustering/singleton/server/src/main/java/org/wildfly/clustering/singleton/server/DistributedSingletonServiceContext.java
@@ -8,8 +8,8 @@ package org.wildfly.clustering.singleton.server;
 import java.util.function.Supplier;
 
 import org.jboss.msc.service.ServiceName;
+import org.wildfly.clustering.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.provider.ServiceProviderRegistry;
-import org.wildfly.clustering.server.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.singleton.SingletonElectionPolicy;
 
 /**
