+++ clustering/web/extension/src/main/java/org/wildfly/extension/clustering/web/session/infinispan/InfinispanSessionManagerFactoryServiceConfigurator.java
@@ -24,6 +24,7 @@ import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceTarget;
+import org.wildfly.clustering.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.ee.Immutability;
 import org.wildfly.clustering.ee.cache.tx.TransactionBatch;
 import org.wildfly.clustering.infinispan.affinity.KeyAffinityServiceFactory;
@@ -34,7 +35,6 @@ import org.wildfly.clustering.infinispan.service.InfinispanRequirement;
 import org.wildfly.clustering.infinispan.service.TemplateConfigurationServiceConfigurator;
 import org.wildfly.clustering.marshalling.spi.ByteBufferMarshaller;
 import org.wildfly.clustering.server.NodeFactory;
-import org.wildfly.clustering.server.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.server.service.ClusteringCacheRequirement;
 import org.wildfly.clustering.server.service.ClusteringRequirement;
 import org.wildfly.clustering.server.service.ProvidedCacheServiceConfigurator;
