+++ connector/src/main/java/org/jboss/as/connector/subsystems/jca/DistributedWorkManagerAdd.java
@@ -39,7 +39,7 @@ import org.jboss.jca.core.workmanager.selector.PingTime;
 import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceTarget;
-import org.wildfly.clustering.server.dispatcher.CommandDispatcherFactory;
+import org.wildfly.clustering.dispatcher.CommandDispatcherFactory;
 import org.wildfly.clustering.server.service.ClusteringDefaultRequirement;
 
 /**
