+++ ee-feature-pack/galleon-shared/src/main/resources/modules/system/layers/base/org/jboss/as/connector/main/module.xml
@@ -61,7 +61,6 @@
         <module name="org.wildfly.clustering.marshalling.jboss"/>
         <module name="org.wildfly.clustering.server.api"/>
         <module name="org.wildfly.clustering.server.service"/>
-        <module name="org.wildfly.clustering.server.spi"/>
         <module name="org.wildfly.clustering.service"/>
         <module name="org.wildfly.common"/>
         <module name="org.glassfish.expressly"/>
