+++ ee-feature-pack/galleon-shared/src/main/resources/modules/system/layers/base/org/jboss/as/ejb3/main/module.xml
@@ -41,7 +41,6 @@
         <module name="org.wildfly.clustering.marshalling.jboss"/>
         <module name="org.wildfly.clustering.server.api"/>
         <module name="org.wildfly.clustering.server.service"/>
-        <module name="org.wildfly.clustering.server.spi"/>
         <module name="org.wildfly.clustering.service"/>
         <module name="org.wildfly.clustering.singleton.api"/>
         <module name="org.wildfly.common"/>
